vertex/requirements.txt
scikit-learn==1.3.2
pandas==2.1.4
google-cloud-bigquery==3.38.0
google-cloud-storage==2.18.2
joblib==1.3.2

